Preload round results with dataloader rather than eagerly
Currently loading a round with dataloader would eagerly preload
the results as well. This is necessary, because some of the round
fields (label, open, finished, active) are computed based on the
results. However, for requests that query a lot of rounds at once,
(specifically recent records), preloading all the results would
generate a lot of unnecessary load, if that query doesn't need to
compute those specific fields. This removes the eager preloading
and uses dataloader directly when computing those fields.