Show PR badge on round results table and dialog
The PR badges were accidentally shown on the results table
previously, but it generated a good response in the short time
that the "feature" was live. This commit officially changes the
round results table and the round result dialog to show the PR
badges when relevant. It does not change the admin results table
(viewed while inserting times) nor does it change the results
projector view, both of which continue to hide the PR badge.