Add litePr flag and use on RoundResultsTable (#127)
When many PRs were broken, the blue PR badges would cover the
round results table. This commit adds an alternative style called
`litePr` which has a background that roughly matches the current
color scheme: light on light mode and dark on dark mode. This new
style is used on the round results table, so that the PRs still
appear on the round results table in a less distracting appearance.
Otherwise, the standard blue badges are used for PR in all other
contexts.