Add warning when submitting new world record
Using the API added in `bf89d0f`, world records are now fetched and
used to check results for new world records when they are submitted.

There were some considersations when altering the warning logic:

1. The purpose of the warning is to have the user go back and
double-check the results. Therefore, it doesn't ultimately matter
exactly which warning gets shown to the user, if more than one
warning is applicable to a set of results.
2. All events have world record singles, but `333mbf` is the only one
without an average. The current code has a branch for `333mbf`
warnings, so before it branches off, I check for a WR single. After the
branch is taken that no longer includes `333mbf`, I check for a new WR
average.
3. Only final results from the WCA database are returned from the query.
If a world record is broken multiple times before results are submitted,
there will be unnecessary warning dialogs. However, since the warning
is simply to detect false positives, it's okay to show the dialog
eagerly.
4. This does not check for Continental Records or National Records, as that
involves more complex logic and data to check the competitor's country
and continent.