Change clustering implementation to dns_cluster
The internal AWS DNS we use for clustering returns up to 8 entries
(randomly, in case there are more entries). The DNS clustering strategy
in libcluster queries the DNS, connects to new nodes and explicitly
disconnects from nodes not present in the DNS. Consequently, if there
are more than 8 nodes, this results in continous disconnects and
disconnects.

This commit changse the clustering implementation to dns_cluster,
which is preferable regardless, because it's a smaller dependency.
It also helps with this specific issues, because it does not involve
explicit disconnects. Having only 8 entries in the DNS is not a
problem, because node connections are transitive, so a new node
connects to the 8 discoverable nodes, and also all the nodes connected
to them.